修复登录时重复创建账户问题。添加编辑按钮

+++ src/main/java/com/hhy/bbs/dto/QuestionDTO.java
@@ -10,7 +10,7 @@ public class QuestionDTO {
     private String title;
     private Long gmtCreate;
     private Long gmtModified;
-    private Long creator;
+    private Integer creator;
     private Integer commentCount;
     private Integer viewCount;
     private Integer likeCount;
